Conflicts:
	AZAppearanceKit/AZLabel.m

+++ AZAppearanceKit/AZToolbar.h
@@ -26,7 +26,7 @@
 /** Specifies the toolbar shadow's opacity.
  
  By default is `0.5`. */
-@property (nonatomic, assign) float shadowOpacity;
+@property (nonatomic, assign) CGFloat shadowOpacity;
 
 /** Specifies the gradient's start color.
